% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cantidades-data.R
\docType{data}
\name{cantidades}
\alias{cantidades}
\title{Cantidades test data}
\format{A data frame.}
\usage{
data(cantidades)
}
\description{
Randomly generated spanish spelled monetary integers to test to_number function
}
\examples{
to_number("mil trescientos noventa y dos")


## testing provided dataframe: cantidades

cantidades$var3 <- lapply(cantidades$var2, to_number)

}
\keyword{datasets}
