% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_number.R
\name{to_number}
\alias{to_number}
\title{translate spanish spelled quantities into their integer counterparts.}
\usage{
to_number(x)
}
\arguments{
\item{x}{A spanish spelled number.}
}
\description{
Allows you to translate to integer numerical words spelled in spanish.
Text must be previously cleaned & removed extraneous words or symbols
Quantities MUST be written in a correct Spanish (this is not a grammar tool)
The upper limit is up to the millions range. Cents must be removed.
}
\examples{
to_number("mil trescientos noventa y dos")

1392

## Example dataframe is provided: cantidades

cantidades$var3 <- lapply(cantidades$var2, to_number)

}
\keyword{currency,}
\keyword{euros}
\keyword{money,}
