% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanish.R
\docType{package}
\name{spanish}
\alias{spanish}
\alias{spanish-package}
\title{spanish: A package for spanish related data functions.}
\description{
The spanish package provides two important functions:
to_number() and geocode_cadastral()
}
\section{to_number()}{

Translate spanish spelled quantities into their integer counterparts.
Allows you to translate to integer numerical words spelled in spanish.
Text must be previously cleaned & removed extraneous words or symbols.
Quantities MUST be written in a correct Spanish (this is not a grammar tool)
The upper limit is up to the millions range. Cents must be removed.
(in my TODO list to parse cents part)
}

