% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_words.R
\name{to_words}
\alias{to_words}
\title{From integers to spanish spelled quantities.}
\arguments{
\item{x}{A valid integer amount.}
}
\value{
A string for the same integer number in spanish.
}
\description{
Takes any integer from zero to millions range and returns spanish characters
}
\keyword{integers,}
\keyword{money,}
\keyword{quantities.}
\keyword{spanish,}
